requirements.txt
langchain==0.2.1
streamlit==1.35.0
unstructured==0.14.2
tiktoken==0.7.0
faiss-cpu==1.8.0
libmagic==1.0
OpenAI == 1.30.2
langchain-community == 0.2.1
langchain-openai == 0.1.7